requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
paho-mqtt==1.6.1
pydantic==1.10.12